+++ src/collective/embeddedpage/browser/embeddedpage.py
@@ -3,6 +3,8 @@ from lxml import etree
 from Products.CMFPlone.utils import safe_unicode
 from Products.Five.browser import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from urlparse import urljoin
+from urlparse import urlparse
 
 import lxml
 import requests
@@ -15,7 +17,9 @@ class EmbeddedPageView(BrowserView):
     def __call__(self):
         resource = self.request.form.get('embeddedpage_get_resource', '')
         if resource:
-            return requests.get(resource).content
+            response = self.request.response
+            response.setHeader('content-type', 'application/javascript')
+            return response.setBody(requests.get(resource).content)
         request_type = self.request['REQUEST_METHOD']
         method = getattr(requests, request_type.lower(), requests.get)
         params = {'url': self.context.url}
@@ -36,7 +40,12 @@ class EmbeddedPageView(BrowserView):
                 continue
             script.attrib['src'] = template.format(
                 self.context.absolute_url(), src)
-        if el.find('body'):
+        for iframe in el.findall('.//iframe'):
+            src = iframe.attrib.get('src', '')
+            if urlparse(src).scheme != '':
+                continue
+            iframe.attrib['src'] = urljoin(self.context.url, src)
+        if el.find('body') is not None:
             el = el.find('body')
         self.embeddedpage = etree.tostring(el, method='html')
         return self.template()